HW3/cloud-function/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-pubsub
google-cloud-logging